# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "subjex",
    "description": "JSON API behind the operator pages — 操作页背后的 JSON 接口",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "operatorBearer": []
    },
    {
      "operatorBasic": []
    }
  ],
  "paths": {
...
    "/api/v1/pages": {
      "get": {
        "tags": [
          "pages-api-endpoint"
        ],
        "operationId": "index",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagesIndexDocument"
                }
              }
            }
          },
          "401": {
            "description": "No operator signed in — 没有已登录的操作员"
          },
          "403": {
            "description": "Operator lacks the permission — 操作员缺少这项权限"
          }
        }
      }
    },
    "/api/v1/pages/{flowKey}": {
      "get": {
        "tags": [
          "pages-api-endpoint"
        ],
        "operationId": "detail",
        "parameters": [
          {
            "name": "flowKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageFlowDocument"
                }
              }
            }
          },
          "401": {
            "description": "No operator signed in — 没有已登录的操作员"
          },
          "403": {
            "description": "Operator lacks the permission — 操作员缺少这项权限"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PageIndexDocument": {
        "type": "object",
        "properties": {
          "flowKey": {
            "type": "string"
          },
          "titleZh": {
            "type": "string"
          },
          "titleEn": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "permission": {
            "type": "string"
          },
          "tenantScoped": {
            "type": "boolean"
          }
        }
      },
      "PagesIndexDocument": {
        "type": "object",
        "properties": {
          "pages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PageIndexDocument"
            }
          }
        }
      },
      "DetailSpecDocument": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          },
          "apiPath": {
            "type": "string"
          },
          "itemsKey": {
            "type": "string"
          },
          "idField": {
            "type": "string"
          },
          "blocks": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional ordered page-block catalog ids; empty/omitted = default layout"
          }
        }
      },
      "ListSpecDocument": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          },
          "apiPath": {
            "type": "string"
          },
          "itemsKey": {
            "type": "string"
          },
          "blocks": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional ordered page-block catalog ids; empty/omitted = default layout"
          }
        }
      },
      "PageFlowDocument": {
        "type": "object",
        "properties": {
          "flowKey": {
            "type": "string"
          },
          "titleZh": {
            "type": "string"
          },
          "titleEn": {
            "type": "string"
          },
          "formKey": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "permission": {
            "type": "string"
          },
          "tenantScoped": {
            "type": "boolean"
          },
          "list": {
            "$ref": "#/components/schemas/ListSpecDocument"
          },
          "detail": {
            "$ref": "#/components/schemas/DetailSpecDocument"
          },
          "submit": {
            "$ref": "#/components/schemas/SubmitSpecDocument"
          }
        }
      },
      "SubmitSpecDocument": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          },
          "apiPath": {
            "type": "string"
          },
          "redirectTo": {
            "type": "string"
          },
          "blocks": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional ordered page-block catalog ids; empty/omitted = default layout"
          }
        }
      },
...
    },
    "securitySchemes": {
      "operatorBasic": {
        "type": "http",
        "description": "Operator login and password (scripts/local tooling) — 操作员口令（脚本/本机工具）",
        "scheme": "basic"
      },
      "operatorBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Opaque access token from POST /api/v1/auth/login — 来自登录接口的不透明访问令牌"
      }
    }
  }
}